Converted to Kotlin Multiplatform project
