+++ gradle.properties
@@ -1,10 +1,2 @@
-groupId=it.czerwinski
-artifactId=kotlin-util
-libVersion=1.4-SNAPSHOT
-libName=Kotlin utilities
-libDescription=Kotlin utility classes based on Scala
-libUrl=https://czerwinski.it/projects/kotlin-util/
-
-kotlinVersion=1.3.61
-dokkaVersion=0.10.1
-junitVersion=4.13
+kotlin.code.style=official
+kotlin.js.generate.executable.default=false
