+++ settings.gradle.kts
@@ -1,0 +1,8 @@
+rootProject.name = "kotlin-util"
+
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+        jcenter()
+    }
+}
